Added Notifications library, enums and models for notification moved to it
