+++ HappyTravel.Edo.Common/HappyTravel.Edo.Common.csproj
@@ -9,4 +9,8 @@
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Models\Notifications\" />
+    </ItemGroup>
+
 </Project>
