+++ HappyTravel.Edo.Data/Notifications/Notification.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
-using HappyTravel.Edo.Common.Enums.Notifications;
-using HappyTravel.Edo.Common.Models.Notifications;
+using HappyTravel.Edo.Notifications.Enums;
+using HappyTravel.Edo.Notifications.Models;
 
 namespace HappyTravel.Edo.Data.Notifications
 {
