+++ HappyTravel.Edo.Data/Notifications/NotificationOptions.cs
@@ -1,5 +1,4 @@
-using System.Collections.Generic;
-using HappyTravel.Edo.Common.Enums.Notifications;
+using HappyTravel.Edo.Notifications.Enums;
 
 namespace HappyTravel.Edo.Data.Notifications
 {
