+++ HappyTravel.Edo.NotificationCenter/Services/Notification/INotificationService.cs
@@ -1,13 +1,12 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using HappyTravel.Edo.Common.Enums.Notifications;
 using HappyTravel.Edo.NotificationCenter.Models;
 
 namespace HappyTravel.Edo.NotificationCenter.Services.Notification
 {
     public interface INotificationService
     {
-        Task Add(Models.Notification notification);
+        Task Add(Notifications.Models.Notification notification);
         Task MarkAsRead(int notificationId);
         Task<List<SlimNotification>> GetNotifications(int userId, int top, int skip);
     }
