+++ HappyTravel.Edo.Notifications/Enums/NotificationTypes.cs
@@ -1,4 +1,4 @@
-namespace HappyTravel.Edo.Common.Enums.Notifications
+namespace HappyTravel.Edo.Notifications.Enums
 {
     public enum NotificationTypes
     {
