+++ HappyTravel.Edo.sln
@@ -11,7 +11,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HappyTravel.Edo.Common", "H
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HappyTravel.Edo.UnitTests", "HappyTravel.Edo.UnitTests\HappyTravel.Edo.UnitTests.csproj", "{64746E40-6B8C-4F9E-9353-2A124C94E81D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HappyTravel.Edo.NotificationCenter", "HappyTravel.Edo.NotificationCenter\HappyTravel.Edo.NotificationCenter.csproj", "{4101AEEF-3B54-4B47-A4A4-426ACBE1EEC7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HappyTravel.Edo.NotificationCenter", "HappyTravel.Edo.NotificationCenter\HappyTravel.Edo.NotificationCenter.csproj", "{4101AEEF-3B54-4B47-A4A4-426ACBE1EEC7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HappyTravel.Edo.Notifications", "HappyTravel.Edo.Notifications\HappyTravel.Edo.Notifications.csproj", "{D0241090-A76F-4AE4-9B79-FF93C1885822}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,6 +41,10 @@ Global
 		{4101AEEF-3B54-4B47-A4A4-426ACBE1EEC7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4101AEEF-3B54-4B47-A4A4-426ACBE1EEC7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4101AEEF-3B54-4B47-A4A4-426ACBE1EEC7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D0241090-A76F-4AE4-9B79-FF93C1885822}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D0241090-A76F-4AE4-9B79-FF93C1885822}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D0241090-A76F-4AE4-9B79-FF93C1885822}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D0241090-A76F-4AE4-9B79-FF93C1885822}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
